Download node_synonymizer inside scripts if needed.
