+++ parser.py
@@ -1,0 +1,63 @@
+import argparse
+import re
+
+
+def kg_version_validator(value: str) -> str:
+    if not re.fullmatch(r"\d+\.\d+\.\d+", value):
+        raise argparse.ArgumentTypeError("KG version must look like X.Y.Z (e.g. 2.10.2)")
+    return value
+
+
+def get_parser():
+    parser = argparse.ArgumentParser()
+
+    parser.add_argument(
+        "--kg-version",
+        type=kg_version_validator,
+        required=True,
+        metavar="VERSION",
+        help="Knowledge graph version (e.g. 2.10.2)",
+    )
+
+    parser.add_argument(
+        "--db-host",
+        default="arax-databases.rtx.ai",
+        type=str,
+        help="Database file host (default: arax-databases.rtx.ai)",
+    )
+
+    parser.add_argument(
+        "--db-username",
+        default="rtxconfig",
+        type=str,
+        help="Database file username (default: rtxconfig)",
+    )
+
+    parser.add_argument(
+        "--db-port",
+        default=22,
+        type=int,
+        help="Database file port (default: 22)",
+    )
+
+    parser.add_argument(
+        "--ssh-key",
+        default=None,
+        help="Path to SSH private key (optional). If omitted, uses SSH agent/default keys.",
+    )
+
+    parser.add_argument(
+        "--ssh-password",
+        default=None,
+        help="SSH password (optional; prefer key/agent). You can also set SSH_PASSWORD env var.",
+    )
+
+    # Optional: choose output dir for downloads
+    parser.add_argument(
+        "--out-dir",
+        default="./data",
+        type=str,
+        help="Where to store downloaded DB files (default: current directory)",
+    )
+
+    return parser
